requirements.txt
pandas 
numpy 
pillow 
opencv-python 
streamlit 
matplotlib
streamlit 
pydeck